update program to use clap derive syntax
this commit makes no changes to core program logic and simply replicates
the functionality of the previous builder syntax code thru the derived
Args struct. this is part of a multi-step effort to move towards
encoding important program information in the type system rather than
thru sketchy pattern matching logic. for now though, it's kind of a
useless change.

+ add dependency on clap_derive
+ add struct Args, which describes the program's command-line arguments
~ create Command instance via Args::command() instead of builder syntax